cache zero net progress

+++ src/datagen.h
@@ -1,6 +1,6 @@
 #ifndef DATAGEN_H
 #define DATAGEN_H
 
-void datagen(const char *directory, int hours);
+void datagen(const char *directory, double hours);
 
 #endif
